move extractOptions from JailReader to helpers (common usage server- / client-side);
